Move get_typescript out of react-native
