+++ crates/ubrn_bindgen/src/bindings/gen_typescript/callback_interface.rs
