+++ crates/ubrn_bindgen/src/bindings/gen_typescript/compounds.rs
