+++ crates/ubrn_bindgen/src/bindings/gen_typescript/custom.rs
