+++ crates/ubrn_bindgen/src/bindings/gen_typescript/enum_.rs
