+++ crates/ubrn_bindgen/src/bindings/gen_typescript/miscellany.rs
