+++ crates/ubrn_bindgen/src/bindings/gen_typescript/object.rs
