+++ crates/ubrn_bindgen/src/bindings/gen_typescript/oracle.rs
