+++ crates/ubrn_bindgen/src/bindings/gen_typescript/primitives.rs
