+++ crates/ubrn_bindgen/src/bindings/gen_typescript/templates/RecordTemplate.ts
