+++ crates/ubrn_bindgen/src/bindings/gen_typescript/templates/TaggedEnumTemplate.ts
