+++ crates/ubrn_bindgen/src/bindings/react_native/gen_cpp/filters.rs
@@ -6,7 +6,7 @@
 use heck::{ToLowerCamelCase, ToUpperCamelCase};
 use uniffi_bindgen::{interface::FfiType, ComponentInterface};
 
-use crate::bindings::react_native::{ComponentInterfaceExt, FfiTypeExt};
+use crate::bindings::extensions::{ComponentInterfaceExt, FfiTypeExt};
 
 pub fn ffi_type_name_from_js(ffi_type: &FfiType) -> Result<String, askama::Error> {
     Ok(match ffi_type {
